Fix validation of a forcedVisitorId value by using unsigned int type.

+++ MatomoTracker/MatomoTracker.swift
@@ -50,7 +50,7 @@ final public class MatomoTracker: NSObject {
         set {
             logger.debug("Setting the forcedVisitorId to \(forcedVisitorId ?? "nil")")
             if let newValue = newValue {
-                let isValidString = Int(newValue, radix: 16) != nil && newValue.count == 16
+                let isValidString = UInt64(newValue, radix: 16) != nil && newValue.count == 16
                 if isValidString {
                     matomoUserDefaults.forcedVisitorId = newValue
                 } else {
